chore: move repeated functions and types to common export files
