+++ src/components/Media.astro
@@ -1,81 +1,7 @@
 ---
 import filePresent from "@uswds-images/usa-icons/file_present.svg";
-import { getMediaUrl } from "@/utilities/media";
-
-export interface MediaValueProps {
-  id: number;
-  altText: string;
-  caption: string | null;
-  _status: "published";
-  reviewReady: boolean;
-  prefix: string;
-  updatedAt: string;
-  createdAt: string;
-  url: string;
-  thumbnailURL: string;
-  filename: string;
-  mimeType: string;
-  filesize: number;
-  width: number;
-  height: number;
-  focalX: number;
-  focalY: number;
-  site: {
-    bucket: string;
-  };
-  sizes: {
-    thumbnail: Record<string, any>;
-    square: Record<string, any>;
-    small: Record<string, any>;
-    medium: Record<string, any>;
-    large: Record<string, any>;
-    xlarge: Record<string, any>;
-    og: Record<string, any>;
-  };
-}
-
-const imageMimeTypes = [
-  "image/jpeg",
-  "image/png",
-  "image/gif",
-  "image/webp",
-  "image/svg+xml",
-  "image/avif",
-  "image/tiff",
-  "image/bmp",
-  "image/x-icon",
-  "image/heic",
-  "image/heif",
-  "image/heif-sequence",
-  "image/heic-sequence",
-];
-
-const tag = (value: string) => `<span class="usa-tag">${value}</span>`;
-
-const formatMimeType = (type: string): string => {
-  const parts = type.split("/");
-  return tag(parts[parts.length - 1].toLowerCase());
-};
-
-const formatBytes = (bytes: number): string => {
-  if (bytes <= 0) return tag("O Bytes");
-
-  const units = [
-    { name: "GB", divisor: 1024 * 1024 * 1024 },
-    { name: "MB", divisor: 1024 * 1024 },
-    { name: "KB", divisor: 1024 },
-  ];
-
-  for (const unit of units) {
-    if (bytes >= unit.divisor) {
-      // Round to 1 decimal places and return with unit
-      return tag(`${(bytes / unit.divisor).toFixed(1)} ${unit.name}`);
-    }
-  }
-
-  // If less than 1 KB, return in bytes
-  return tag(`${bytes} Bytes`);
-};
+import { getMediaUrl, imageMimeTypes, tag, formatMimeType, formatBytes } from "@/utilities/media";
+import type { MediaValueProps } from "@/env";
 
 interface Props {
   src?: string;
